pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cassetta-attrezzi"
version = "0.1.0"
description = "Utility per connessione e gestione dati (Oracle, pandas, dotenv)"
authors = [
  { name="Tuo Nome", email="email@example.com" }
]
dependencies = [
  "oracledb",
  "pandas",
  "python-dotenv"
]
requires-python = ">=3.9"

[tool.setuptools.packages.find]
where = ["src"]